+++ packages/ui-core/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+- **Carousel**
+    - add zero padding for slides
+    - add sm padding for slides that make center position of slider
 
 ## Release 1.1.0
 - **Carousel**